Move driver.ts to /src
